chore(authz): rename policies
