+++ packages/shared/src/authz/policies/projects/index.ts
@@ -1,0 +1,2 @@
+export { createProjectPolicyFactory } from './canCreate.js'
+export { queryProjectPolicyFactory } from './canQuery.js'
